refactor: separate detailLayout to be module
