+++ src/routes/Chart/RealtimePrice.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function RealtimePrice() {
+  return (
+    <div>RealtimePrice</div>
+  )
+}
